Separate responsibility to files
