+++ src/test/letter-tree-test.py
@@ -1,9 +1,10 @@
-from main import *
+import utils
+from letterbranch import LetterBranch
 
 def test_get_word_simple():
-    phrase_dict, phrase_len = phrase_to_dict("poultry outwits ants")
+    phrase_dict, phrase_len = utils.phrase_to_dict("poultry outwits ants")
 
-    tree, words = parse_words(phrase_dict, "data/wordlist")
+    tree, words = LetterBranch.parse_words(phrase_dict, "data/wordlist")
     a_leaf = tree.children['a']
 
     assert a_leaf.is_word == True
@@ -11,9 +12,9 @@ def test_get_word_simple():
     assert a_str == 'a'
 
 def test_get_word_long():
-    phrase_dict, phrase_len = phrase_to_dict("poultry outwits ants")
+    phrase_dict, phrase_len = utils.phrase_to_dict("poultry outwits ants")
 
-    tree, words = parse_words(phrase_dict, "data/wordlist")
+    tree, words = LetterBranch.parse_words(phrase_dict, "data/wordlist")
     tail_leaf = tree.children['t'].children['a'].children['i'].children['l']
 
     assert tail_leaf.is_word == True
@@ -22,12 +23,13 @@ def test_get_word_long():
 
 
 def test_returned_words():
-    phrase_dict, phrase_len = phrase_to_dict("poultry outwits anpts delamgrovesa")
-    tree, words = parse_words(phrase_dict, "data/sample")
+    phrase_dict, phrase_len = utils.phrase_to_dict("poultry outwits anpts delamgrovesa")
+    tree, words = LetterBranch.parse_words(phrase_dict, "data/sample")
 
     app_leaf = tree.children['a'].children['p'].children['p']
     assert app_leaf.is_word == True
     app_str = str(app_leaf)
     assert app_str == 'app'
 
-    assert len(words) == 5
+    print(len(words))
+    assert len(words) == 8
