+++ src/test/utils-test.py
@@ -1,7 +1,7 @@
-from main import *
+import utils
 
 def test_dict_to_str():
     dict1 = {'p':1, 't':4, 'a':0, 'i':2}
 
-    dict_str = dict_to_str(dict1)
+    dict_str = utils.dict_to_str(dict1)
     assert dict_str == "pttttii"
